[update] analyzer module converted to folder
