+++ brkraw/api/analyzer/__init__.py
@@ -1,0 +1,6 @@
+from .base import BaseAnalyzer
+from .scaninfo import ScanInfoAnalyzer
+from .affine import AffineAnalyzer
+from .dataarray import DataArrayAnalyzer
+
+__all__ = [BaseAnalyzer, ScanInfoAnalyzer, AffineAnalyzer, DataArrayAnalyzer]
