+++ brkraw/api/analyzer/base.py
@@ -1,0 +1,3 @@
+class BaseAnalyzer:
+    def vars(self):
+        return self.__dict__
